Отправляю информацию о сериале пользователю
Отправка сообщения с информацией о следующей серии
Добавление сериала в базу из телеграмма
