+++ bot_telegramm/help.txt
@@ -1,3 +1,3 @@
 /echo - Переслать сообщение
 /add - Добавить сериал в базу
-/show - Показать список сериалов
+/show_serial - Показать список сериалов
